The event buffer now allocates a Box<[u8]> instead of copying all 255 bytes of the octet strings.
#54
